Ряд фиксов: 1. Функция распределения строится по интервалам 2. Исправлена ошибка в первом ряду функции распределения 3. В таблицах числа не округляются, а используется формат для ToString 4. Валидация для ввода интервального ряда

+++ StatisticDistribution/Form1.cs
@@ -134,8 +134,8 @@ namespace StatisticDistribution
 		//Валидация ввода интервального ряда
 		private void GridIntervalData_CellValidating(object sender, DataGridViewCellValidatingEventArgs e)
 		{
-			double res;
-			if((e.ColumnIndex==1) && !double.TryParse((string)e.FormattedValue, out res))
+			int res;
+			if((e.ColumnIndex==1) && !int.TryParse((string)e.FormattedValue, out res))
 			{
 				e.Cancel = true;
 			}
